converted int to float

+++ src/main/java/com/increff/employee/model/OrderItemData.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class OrderItemData extends OrderItemForm{
     private String barcode;
-    private int price;
+    private Float price;
     private int quantity;
 
     private int id;
